Avoid expanding submenus
The sidebar menu shows the entire site tree for the selected
top level submenu. For sites with a lot of subsites and pages,
this makes it much harder to navigate.

This adds a `hideExtendedTree` site parameters. When set to
true, the side menu shows only the most relevant pages: the
current page, its ancestors, and their children.